+++ lib/flutter_wot_ui.dart
@@ -14,6 +14,7 @@ export 'src/theme/wot_scheme.dart';
 export 'src/theme/wot_theme_data.dart';
 export 'src/theme/wot_theme.dart';
 export 'src/theme/wot_state.dart';
+export 'src/locale/wot_messages.dart';
 export 'src/icon/iconfont_map.dart';
 export 'src/util/wot_util.dart';
 export 'src/util/props.dart';